Support `require 'cgi/escape'` with extracting CGI::Escape from CGI::Util
