+++ lib/cgi.rb
@@ -294,4 +294,5 @@ end
 require 'cgi/core'
 require 'cgi/cookie'
 require 'cgi/util'
+require 'cgi/escape'
 CGI.autoload(:HtmlExtension, 'cgi/html')
